refactor: replace LoggerProvider with headless LoggerSynchronizer
