+++ frontend/src/App.tsx
@@ -18,7 +18,7 @@ import SuspenseBoundary from '@web/components/SuspenseBoundary/SuspenseBoundary'
 import WaitingRoomSkeleton from '@pages/WaitingRoom/WaitingRoom.skeleton';
 import MeetingRoomSkeleton from '@pages/MeetingRoom/MeetingRoom.skeleton';
 import SessionProvider from '@Context/SessionProvider/session';
-import LoggerProvider from '@Context/LoggerProvider';
+import LoggerSynchronizer from '@Context/LoggerProvider';
 import ErrorBoundary from './components/ErrorBoundary';
 import EnvGuard from './components/EnvGuard';
 import { ErrorPage } from './pages/ErrorBoundary';
@@ -70,11 +70,10 @@ const InnerApp = () => {
                   <SuspenseBoundary fallback={<MeetingRoomSkeleton />}>
                     <RoomProvider>
                       <SessionProvider>
-                        <LoggerProvider>
-                          <PublisherProvider>
-                            <MeetingRoom />
-                          </PublisherProvider>
-                        </LoggerProvider>
+                        <LoggerSynchronizer />
+                        <PublisherProvider>
+                          <MeetingRoom />
+                        </PublisherProvider>
                       </SessionProvider>
                     </RoomProvider>
                   </SuspenseBoundary>
