+++ frontend/src/Context/LoggerProvider/LoggerSynchronizer.tsx
@@ -1,15 +1,9 @@
 import { useEffect } from 'react';
 import useSessionContext from '@hooks/useSessionContext';
 import useUserContext from '@hooks/useUserContext';
-import frontendLogger from '../../../logger';
+import frontendLogger from '../../logger';
 
-/**
- * Syncs userId, sessionId, and connectionId from UserContext and SessionContext
- * into the global frontendLogger on every change.
- * When the session ends, vonageVideoClient becomes null and sessionId/connectionId
- * naturally become undefined, so no explicit clearContext is needed.
- */
-const useLoggerContext = () => {
+const LoggerSynchronizer = () => {
   const { vonageVideoClient } = useSessionContext();
   const { user } = useUserContext();
 
@@ -21,7 +15,7 @@ const useLoggerContext = () => {
     frontendLogger.setContext({ userId, sessionId, connectionId });
   }, [userId, sessionId, connectionId]);
 
-  return { userId, sessionId, connectionId };
+  return null;
 };
 
-export default useLoggerContext;
+export default LoggerSynchronizer;
