--- frontend/src/Context/LoggerProvider/useLoggerContext/index.tsx
@@ -1,1 +1,0 @@
-export { default } from './useLoggerContext';
