+++ frontend/src/test/providers/makeTestProvider.ts
@@ -1,7 +1,6 @@
 import composeProviders, { type ProviderComponent } from '@web/helpers/composeProviders';
 import {
   makeBackgroundPublisherProviderWrapper,
-  makeLoggerProviderWrapper,
   makePreviewPublisherProviderWrapper,
   makePublisherProviderWrapper,
   makeSessionProviderWrapper,
@@ -17,7 +16,6 @@ export enum providers {
   publisher = 'publisher',
   backgroundPublisher = 'backgroundPublisher',
   previewPublisher = 'previewPublisher',
-  logger = 'logger',
 }
 
 /**
@@ -29,7 +27,6 @@ const MAKERS = {
   [providers.publisher]: makePublisherProviderWrapper,
   [providers.backgroundPublisher]: makeBackgroundPublisherProviderWrapper,
   [providers.previewPublisher]: makePreviewPublisherProviderWrapper,
-  [providers.logger]: makeLoggerProviderWrapper,
 } as const;
 
 type ProvidersMakers = typeof MAKERS;
@@ -43,7 +40,6 @@ const PROVIDER_DEPENDENCIES = {
   [providers.publisher]: [providers.user, providers.session],
   [providers.backgroundPublisher]: [providers.user, providers.session, providers.publisher],
   [providers.previewPublisher]: [providers.user],
-  [providers.logger]: [providers.user, providers.session],
 } as const;
 
 /**
