+++ frontend/src/test/providers/makersIndex.ts
@@ -22,8 +22,3 @@ export {
   default as makeBackgroundPublisherProviderWrapper,
   type BackgroundPublisherProviderWrapperOptions,
 } from './makeBackgroundPublisherProviderWrapper';
-
-export {
-  default as makeLoggerProviderWrapper,
-  type LoggerProviderWrapperOptions,
-} from './makeLoggerProviderWrapper';
